requirements.txt
beautifulsoup4
requests
transformers
nltk
fastapi
uvicorn
gradio
huggingface_hub
pandas
numpy
tqdm
python-dotenv
TTS 
soundfile 
deep-translator
torch
-e .